Allow for restarting (#26)
* ENH: Support restarting from a previous chain

A new calculation can be restarted from a previous chain.
As a side effect, the chain we output to file is the real chain and not the flat chain.

* ENH: Raise if tracefile/probfile will be overwritten

Because MCMC fitting is relatively long-running and RAM intensive, it is ideal to be able to restart from a previous calculation.

Currently, to restart, a user would manually update the parameters in the database to the parameters with the lowest probability from the previous run, then use that set of parameters to initialize a new set of walkers.

Since we save the state of the chains as the tracefile anyways, we can use the chains to restart a run in emcee and avoid tuning the walker initialization parameters on restart.

To make this (and analysis) easier, chains and log probabilities are now written out as binaries, rather than text files. This preserves the shape with the number of chains, so the chain file is now of shape (k, n, p), where k is number of walkers, n is the preallocated number of iterations, p is the number of parameters. The previous shape was (kn, p). The same is true for the log probability file, which is now of shape (k, n) rather than (kn).

This change also introduces a command line argument --restart which takes the path to a chain file. If the name of the input chain conflicts with the default name, an error will be raised.