ENH: Improve logging
Raise error: Bad things happened and we couldn't recover or handle it.
Warnings: Bad things happened that the user should know and/or possibly fix, but we recovered from it.
Info: Feedback of the progress of the program. Usually report at the level of user settings, e.g. which runs are occuring, what phases are being fit, etc.
Debug: Possibly useful information, but usually beyond the user's control. Examples are the values of initial parameters, when the chain is written to file, which endmember parameters are currently being generated