ENH/FIX: Add support for pycalphad Databases with Species
This a backwards compatible fix for pycalphad Databases that have
Species objects. The support for parsing and using Species in pycalphad
Databases broke the support for writing ESPEI-generated databases because
we added elements to the elements dictionary only, where we needed to add
elements and Species.

Real species support in ESPEI still needs to be done, where we can
consider them in phases and fit their values, but this will not be done
until pycalphad fully supports species.