MAINT: Improve input consistency
Input and internal ESPEI APIs were often confused whether to call the MCMC trace of chains, iterations and parameters a "chain" or a "trace".

This commit makes it so that every time we are talking about all the chains together, we refer to it as the "trace", but when we refer to an individual chain the trace, we call it a "chain".

Additionally, the "mcmc.mcmc_steps" was renamed to "mcmc.iterations" to avoid redundancy and improve clarity. "mcmc.mcmc_save_interval" was similarly updated to "save_interval".

The input file, docs and Python API all reflect these changes.