ENH: Remove line and remove HM_FORM plot in plot_parameters (#75)
Currently, there are lines that connect data when there are more than 10 data points in a dataset. This is kind of distracting and not very useful over just data points, so it's removed.

Plotting HM_FORM is redundant with HM_MIX (which plots formation data) and can't plot the actual mixing data (because we don't have a way to specify a reference state). So it's removed.