Use unpack_components for getting data, rather than mapping v.Species
Mapping v.Species could be dangerous since the components likely do not
include SPECIES and only pure elements. We should use unpack_components

It appears that we don't *need* it, almost out of luck, because all the
current consumers of species use unpack_components internally in the
important places, but this is updated no less.