ENH: Allow for disabling datasets using JSON (#153)
This PR aims to allows JSON datasets to be marked as disabled by using the `disabled: True` at the top level of JSON datasets. 

Until now, datasets were removed as fitting candidates by not having a name ending in `.json` or by moving the dataset files from the directory holding the desired input data. Disabling datasets using a key/value inside the datasets means that the dataset (and YAML, via tags) can semantically describe whether datasets are enabled or not without knowing the context of the path of the file or the filename of the original data (if either exist).

This PR allows users to use `load_datasets` to load disabled datasets (off by default). If using ESPEI from the command line or via the `espei_script.run_espei` function, disabled datasets will also be removed after loading datasets and applying tags. This means that users can apply tags to easily switch data on/off from the YAML file.