FIX: Fix performance of _sample_solution_constitution (#174)
`_sample_solution_constitution` was incorrectly using the independent site fractions that were already sampled as the expression for `sympy.lambdify`. In other words, `lambdify` was internally creating a loop over `grid_1d.size` constant valued "expressions", which would just be evaluated to the constant values, no matter the input. This could be really slow (~20 seconds for a `grid_1d.size` of 40,000).

This PR fixes the performance by checking that the expression is indeed a dependent site fraction expression rather than an independent sample that we already took.

Thanks to @wahab2604 for reporting this issue.