MAINT: Refactor `plot_parameters` into `plot_interaction` and `plot_endmember` (#177)
The overall functionality of `plot_parameters` is now contained in `plot_interaction` and `plot_endmember` which are more clear about the form. `plot_parameters` is deprecated and set to be removed in ESPEI 0.9. The design of these now plots one property at a time on given axes. Both functions expose options to customize how matplotlib plots predicted values and observed values. Closes #60.

The new API is superior because there's more fine control over which properties are plotted, regardless of the existing data and whether `require_data` is passed or not.

To recover the full functionality of `plot_parameters` where multiple plots are in the same figure, users should use matplotlib subplots.

Bar charts comparing data to parameters are no longer possible because they are an uncommon plot type and not usually seen in publications. They are not much more informative than a line plot.