ENH: Better heuristic starting points for constrained driving force equilibrium (#198)
Restrict the single phase candidate points by only choosing points that are within specified (currently hardcoded) mass residual from the target composition. This should make the starting points better (closer to the target composition, which may be in a miscibility gap and relatively high in energy) and give fewer points to check to find the starting point in the first place.