ENH: improve AICc formulation (#205)
Improve the implementation of the AICc by better handling the case where the denominator in the correction term can become negative (small number of samples and a large number of parameters). With these changes, there is a smooth connection in the AICc rather than a discontinuity and a negative contribution to the correction term. In the spirit of the AICc, the AICc continues to increase with an increasing number of parameters.