ENH: Allow entering the `__HYPERPLANE__` phase in ZPF vertices to control the compositions for the target hyperplane (#221)
With this PR, one can define a phase region in a dataset as follows:

    [["__HYPERPLANE__", ["CU"], [0.001015]], ["HCP_A3", ["CU"], [null]], ["CUMG2", ["CU"], [null]]]

which will result in estimating the target hyperplane at X(CU) = 0.001015 and the driving forces to that hyperplane for HCP_A3 and CUMG2 (in the same way as they are estimated now).

This allows users to more finely control how the target hyperplane is generated and allows the target hyperplane to be specified independently of any phases/phase compositions that are known. Multiple vertices with the `__HYPERPLANE__` phase can be entered. If a `__HYPERPLANE__` vertex is not present (i.e. all existing datasets), the current rules apply: each phase with a fully specified composition (no `null` entries) contributes to the target hyperplane at that composition.