ENH/MAINT: Refactor residual/likelihood functions (#236)
This PR introduces a `ResidualFunction` class to abstract the concept of residual/likelihood functions for different types of data. This allows new residual functions to be defined (both internally and externally) and registered for dynamic use. The `ResidualFunction` class also aims to encapsulate the concept of a likelihood function in terms of what ESPEI can provide as inputs (a database, datasets and related metadata) and outputs via calling conventions, which should make it clearer how to add new likelihood contributions and what functionality needs to be tested.

Other changes include:
- priors can now be optionally provided to the `EmceeOptimizer.predict` method, to better match the API. The default is now the improper zero prior, which is backwards compatible
- introduce an `espei.typing` module that currently holds some type aliases that improve semantics, especially since so many things are simply strings
- introduce [pydantic](https://pydantic-docs.helpmanual.io/) as a dependency to be able to enable parsing of phase models (and, in the future, datasets) into objects.