Add modified version of _get_sample_condition_dicts
The idea of the modified version is that we also compute the actual site
fractions because individual site fractions are not currently handled by
ESPEI, but can slip in from existing models if not using a reference
state where those contributions cancel (e.g. no _MIX or _FORM refstates
keep the unary extrapolation). to do this, we'll use the config tuple
and create site fractions from the points dict.

tests currently pass locally