FIX: Fix Residual protocol for MCMC for Python 3.11 (#253)
It seems that Python 3.11 changed the behavior of calling `super().__init__` for `Protocol` subclasses that have `Optional` arguments and subclasses of the `Residual` protocol are raising. 

This PR adds Python 3.11 to the CI test matrix and fixes the protocol signature (making the optional arguments actually optional) and also updates the callers that had incorrect signatures.