FIX: fix dataplot for ternary cases with __HYPERPLANE__ (#252)
The presence of `__HYPERPLANE__` overall compositions in datasets was causing the dimensionality of tie-regions to be artificially increased. For example, a 3 phase region with an additional `__HYPERPLANE__` overall composition was being categorized as a 4 phase region and not plotted. This did not affect binary dataplotting because we did not differently depending on the number of phases in tie-regions (in ternary dataplotting we draw three phase regions as red tie-triangles).